+++ app/models/CronjobModel.php
@@ -13,7 +13,7 @@ class CronjobModel extends Controller {
 
     public function __construct()
     {
-        $this->db = new Database;
+        $this->db = Database::instance();
     }
 
     /**